+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - BREAKING: Some imports have been moved to modules. Use `ext4plus::prelude` if you want to keep import compatibility.
 - Fix infinite loop in block allocation logic
+- Fix large block map write failing
 
 ## 0.1.0-alpha.4